requirements.txt
psycopg2-binary==2.8.1
architect
ipython
tqdm
factory_boy==2.8.1
Faker
